refactor: :truck: move preference entity to common
