+++ src/auth/service/auth.service.ts
@@ -16,7 +16,7 @@ import JWT from '../../common/utils/jwt';
 import { EMAIL } from '../../common/enum/email.enum';
 import { TTL } from '../../common/enum/token.enum';
 
-import Preference from '../../user/entity/preference.entity';
+import Preference from '../../common/entity/preference.entity';
 import Location from '../../location/entity/location.entity';
 
 import User from '../entity/user.entity';
