+++ src/user/entity/user.details.entity.ts
@@ -1,6 +1,6 @@
 import EventDetails from '../../event/entity/event.details.entity';
 import Location from '../../location/entity/location.entity';
-import Preference from '../../user/entity/preference.entity';
+import Preference from '../../common/entity/preference.entity';
 import Image from '../../image/entity/image.entity';
 
 class UserDetails {
